More work on tests
Settings class to manage run dir

+++ test/Resque/Test.php
@@ -2,10 +2,15 @@
 
 namespace Resque;
 
-use Resque\Test\ClientFactory;
+use Resque\Test\Settings;
 
 abstract class Test extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * @var Settings
+     */
+    protected static $settings = null;
+
     /**
      * @var Resque
      */
@@ -16,11 +21,18 @@ abstract class Test extends \PHPUnit_Framework_TestCase
      */
     protected $redis;
 
+    public static function setSettings(Settings $settings)
+    {
+        self::$settings = $settings;
+    }
+
     public function setUp()
     {
-        $factory = new ClientFactory();
+        if (!self::$settings) {
+            throw new \LogicException('You must supply the test case with a settings instance');
+        }
 
-        $this->redis = $factory->get();
+        $this->redis = self::$settings->getClient();
         $this->redis->flushdb();
 
         $this->resque = new Resque($this->redis);
